refactor benching scripts for more accurate comparisons
