+++ .gitignore
@@ -12,3 +12,4 @@ convert.js
 npm-debug.log
 shave.js
 
+viz/index.html
