+++ bench/node-mapnik.bench.js
@@ -18,7 +18,7 @@ const path = require('path');
 const assert = require('assert');
 const Queue = require('d3-queue').queue;
 const vtquery = require('../lib/index.js');
-const rules = require('./mapnik-rules');
+const rules = require('./rules');
 let ruleCount = 1;
 const mapnik = require('mapnik');
 
@@ -53,18 +53,35 @@ function runRule(rule, ruleCallback) {
     if (rule.tiles.length > 1){
       var target_vt = new mapnik.VectorTile(rule.zxy.z, rule.zxy.x, rule.zxy.y);
       var vt1 = new mapnik.VectorTile(rule.tiles[0].z,rule.tiles[0].x,rule.tiles[0].y);
+      vt1.bufferSize = rule.options.buffer_size; 
       vt1.addDataSync(rule.tiles[0].buffer);
       var vt2 = new mapnik.VectorTile(rule.tiles[1].z,rule.tiles[1].x,rule.tiles[1].y);
+      vt2.bufferSize = rule.options.buffer_size; 
       vt2.addDataSync(rule.tiles[1].buffer);
+
       // http://mapnik.org/documentation/node-mapnik/3.6/#VectorTile.composite
       target_vt.composite([vt1, vt2], {}, function(err, result) {
         if (err) {
           return cb(err);
         }
 
+        let options = {compression:'none'}
         if (rule.options.compress){
-          result.getData({compression:'gzip'}, function(err, data) {
+          options.compression = 'gzip'; 
+        }
+
+        result.getData(options, function(err, data) {
+            if (rule.options.compress){
+              if(data[0] !== 0x1F && data[1] !== 0x8B){
+                throw new Error('resulting buffer is not compressed!');
+              }
+            } 
             ++runs;
+
+            if (err) {
+              throw err;
+            }
+
             if (track_mem && runs % 1000) {
               var mem = process.memoryUsage();
               if (mem.rss > memstats.max_rss) memstats.max_rss = mem.rss;
@@ -72,17 +89,7 @@ function runRule(rule, ruleCallback) {
               if (mem.heapUsed > memstats.max_heap) memstats.max_heap = mem.heapUsed;
             }
             return cb();
-          }); 
-        }else{
-          ++runs;
-          if (track_mem && runs % 1000) {
-            var mem = process.memoryUsage();
-            if (mem.rss > memstats.max_rss) memstats.max_rss = mem.rss;
-            if (mem.heapTotal > memstats.max_heap_total) memstats.max_heap_total = mem.heapTotal;
-            if (mem.heapUsed > memstats.max_heap) memstats.max_heap = mem.heapUsed;
-          }
-          return cb();
-        }
+        }); 
       });
     } else {
       return cb();
