+++ bench/rules.js
@@ -15,7 +15,7 @@ module.exports = [
   // },
   {
     description: 'two different tiles at the same zoom level, zero buffer',
-    options: { },
+    options: {buffer_size: 128},
     tiles: [
       { z: 15, x: 5239, y: 12666, buffer: getTile('sanfrancisco', '15-5239-12666.mvt')},
       { z: 15, x: 5239, y: 12666, buffer: getTile('osm-qa-astana', '12-2861-1368.mvt')}
@@ -24,7 +24,7 @@ module.exports = [
   },
   {
     description: 'two different tiles from different zoom levels (separated by one zoom), zero buffer',
-    options: { },
+    options: {buffer_size: 128 },
     tiles: [
       { z: 0, x: 0, y: 0, buffer: getTile('sanfrancisco', '15-5239-12666.mvt')},
       { z: 1, x: 0, y: 0, buffer: getTile('osm-qa-astana', '12-2861-1368.mvt')}
@@ -33,7 +33,7 @@ module.exports = [
   },
   {
     description: 'two different tiles from different zoom levels (separated by more than one zoom), zero buffer',
-    options: { },
+    options: {buffer_size: 128 },
     tiles: [
       { z: 0, x: 0, y: 0, buffer: getTile('sanfrancisco', '15-5239-12666.mvt')},
       { z: 10, x: 0, y: 0, buffer: getTile('osm-qa-astana', '12-2861-1368.mvt')}
@@ -42,7 +42,7 @@ module.exports = [
   },
   {
     description: 'tiles completely made of points, overzooming, no properties',
-    options: { },
+    options: {buffer_size: 128 },
     tiles: [
       { z: 0, x: 0, y: 0, buffer:  fs.readFileSync('./test/fixtures/points-16-10498-22872.mvt')},
       { z: 1, x: 0, y: 0, buffer: fs.readFileSync('./test/fixtures/points-16-10498-22872.mvt')}
@@ -51,7 +51,7 @@ module.exports = [
   },
   {
     description: 'tiles completely made of points, same zoom, no properties',
-    options: { },
+    options: {buffer_size: 128 },
     tiles: [
       { z: 0, x: 0, y: 0, buffer:  fs.readFileSync('./test/fixtures/points-16-10498-22872.mvt')},
       { z: 0, x: 0, y: 0, buffer: fs.readFileSync('./test/fixtures/points-16-10498-22872.mvt')}
@@ -60,7 +60,7 @@ module.exports = [
   },
   {
     description: 'tiles completely made of points, overzoooming, lots of properties',
-    options: { },
+    options: {buffer_size: 128 },
     tiles: [
       { z: 0, x: 0, y: 0, buffer:  fs.readFileSync('./test/fixtures/points-properties-16-10498-22872.mvt')},
       { z: 1, x: 0, y: 0, buffer: fs.readFileSync('./test/fixtures/points-properties-16-10498-22872.mvt')}
@@ -69,7 +69,7 @@ module.exports = [
   },
   {
     description: 'tiles completely made of points, same zoom, lots of properties',
-    options: { },
+    options: {buffer_size: 128 },
     tiles: [
       { z: 0, x: 0, y: 0, buffer:  fs.readFileSync('./test/fixtures/points-properties-16-10498-22872.mvt')},
       { z: 0, x: 0, y: 0, buffer: fs.readFileSync('./test/fixtures/points-properties-16-10498-22872.mvt')}
@@ -87,7 +87,7 @@ module.exports = [
   },
   {
     description: 'tiles completely made of linestrings, overzooming and lots of properties',
-    options: { },
+    options: {buffer_size: 128 },
     tiles: [
       { z: 0, x: 0, y: 0, buffer:  fs.readFileSync('./test/fixtures/linestrings-properties-16-10498-22872.mvt')},
       { z: 0, x: 0, y: 0, buffer: fs.readFileSync('./test/fixtures/linestrings-properties-16-10498-22872.mvt')}
@@ -96,7 +96,7 @@ module.exports = [
   },
   {
     description: 'tiles completely made of polygons, overzooming and lots of properties',
-    options: { },
+    options: {buffer_size: 128 },
     tiles: [
       { z: 0, x: 0, y: 0, buffer:  fs.readFileSync('./test/fixtures/polygons-properties-16-10498-22872.mvt')},
       { z: 0, x: 0, y: 0, buffer: fs.readFileSync('./test/fixtures/polygons-properties-16-10498-22872.mvt')}
@@ -105,7 +105,7 @@ module.exports = [
   }, 
   { 
     description: 'tiles completely made of points and linestrings, overzooming and lots of properties',
-    options: { },
+    options: {buffer_size: 128 },
     tiles: [
       { z: 15, x: 5239, y: 12666, buffer:  fs.readFileSync('./test/fixtures/points-poi-sf-15-5239-12666.mvt')},
       { z: 15, x: 5239, y: 12666, buffer: fs.readFileSync('./test/fixtures/linestrings-sf-15-5239-12666.mvt')}
@@ -114,7 +114,7 @@ module.exports = [
   },  
   { 
     description: 'returns compressed buffer - tiles completely made of points and linestrings, overzooming and lots of properties',
-    options: { compress:true },
+    options: { compress:true, buffer_size: 128 },
     tiles: [
       { z: 15, x: 5239, y: 12666, buffer:  fs.readFileSync('./test/fixtures/points-poi-sf-15-5239-12666.mvt')},
       { z: 15, x: 5239, y: 12666, buffer: fs.readFileSync('./test/fixtures/linestrings-sf-15-5239-12666.mvt')}
@@ -132,7 +132,7 @@ module.exports = [
   }, 
   { 
     description: 'tiles completely made of points and linestrings, overzooming (2x) and lots of properties',
-    options: { },
+    options: {buffer_size: 128 },
     tiles: [
       { z: 15, x: 5239, y: 12666, buffer:  fs.readFileSync('./test/fixtures/points-poi-sf-15-5239-12666.mvt')},
       { z: 15, x: 5239, y: 12666, buffer: fs.readFileSync('./test/fixtures/linestrings-sf-15-5239-12666.mvt')}
@@ -141,7 +141,7 @@ module.exports = [
   }, 
   { 
     description: 'tiles completely made of polygons, overzooming and lots of properties',
-    options: { },
+    options: { buffer_size: 128},
     tiles: [
       { z: 15, x: 5239, y: 12666, buffer:  fs.readFileSync('./test/fixtures/polygons-buildings-sf-15-5239-12666.mvt')},
       { z: 15, x: 5239, y: 12666, buffer: fs.readFileSync('./test/fixtures/polygons-hillshade-sf-15-5239-12666.mvt')}
@@ -150,7 +150,7 @@ module.exports = [
   }, 
   { 
     description: 'tiles completely made of polygons, overzooming (2x) and lots of properties',
-    options: { },
+    options: {buffer_size: 128 },
     tiles: [
       { z: 15, x: 5239, y: 12666, buffer:  fs.readFileSync('./test/fixtures/polygons-buildings-sf-15-5239-12666.mvt')},
       { z: 15, x: 5239, y: 12666, buffer: fs.readFileSync('./test/fixtures/polygons-hillshade-sf-15-5239-12666.mvt')}
@@ -159,17 +159,16 @@ module.exports = [
   }, 
   { 
     description: 'return compressed buffer - tiles completely made of polygons, overzooming (2x) and lots of properties',
-    options: {compress:true},
+    options: {compress:true, buffer_size: 128},
     tiles: [
       { z: 15, x: 5239, y: 12666, buffer:  fs.readFileSync('./test/fixtures/polygons-buildings-sf-15-5239-12666.mvt')},
       { z: 15, x: 5239, y: 12666, buffer: fs.readFileSync('./test/fixtures/polygons-hillshade-sf-15-5239-12666.mvt')}
     ], 
     zxy: { z: 17, x: 20956, y: 50664}
   }, 
-  , 
   { 
-    description: 'buffer_size 128 - tiles completely made of polygons, overzooming (2x) and lots of properties',
-    options: {buffer_size: 128},
+    description: 'buffer_size 4096 - tiles completely made of polygons, overzooming (2x) and lots of properties',
+    options: {buffer_size: 4096},
     tiles: [
       { z: 15, x: 5239, y: 12666, buffer:  fs.readFileSync('./test/fixtures/polygons-buildings-sf-15-5239-12666.mvt')},
       { z: 15, x: 5239, y: 12666, buffer: fs.readFileSync('./test/fixtures/polygons-hillshade-sf-15-5239-12666.mvt')}
